move arg tests to tsts folder
